+++ src/sessionTypes/machineAuth.js
@@ -47,6 +47,21 @@ class MachineAuth {
   }
 
   /**
+   * Removes an audience's API token from the in-memory token storage
+   *
+   * @param audienceName
+   *
+   * @returns {Promise}
+   */
+  clearCurrentApiToken(audienceName) {
+    const promise = this._tokenPromises[audienceName] || Promise.resolve();
+
+    return promise.then(() => {
+      delete this._sessionInfo[audienceName];
+    });
+  }
+
+  /**
    * Gets the current API token (used to communicate with other Contxt APIs). Will get and store a
    * token or use a previously acquired and stored token.
    *